Revert "⚗[REPLAY-341] Add VisualViewport tracking (Pinch Zoom) (#1089)" (#1115)
This reverts commit 4cb949cbeee0420a0bd52cc7198a3c40ace77f25.
