+++ packages/rum/src/domain/record/types.ts
@@ -1,5 +1,5 @@
 import { DefaultPrivacyLevel } from '@datadog/browser-core'
-import { FocusRecord, RawRecord, VisualViewportRecord } from '../../types'
+import { FocusRecord, RawRecord } from '../../types'
 import { MutationController } from './mutationObserver'
 
 export enum IncrementalSource {
@@ -79,7 +79,6 @@ export interface ObserverParam {
   mouseInteractionCb: MouseInteractionCallBack
   scrollCb: ScrollCallback
   viewportResizeCb: ViewportResizeCallback
-  visualViewportResizeCb: VisualViewportResizeCallback
   inputCb: InputCallback
   mediaInteractionCb: MediaInteractionCallback
   styleSheetRuleCb: StyleSheetRuleCallback
@@ -182,7 +181,7 @@ export const MouseInteractions = {
 
 export type MouseInteractions = typeof MouseInteractions[keyof typeof MouseInteractions]
 
-export interface MouseInteractionParam {
+interface MouseInteractionParam {
   type: MouseInteractions
   id: number
   x: number
@@ -242,8 +241,6 @@ export interface MediaInteractionParam {
 export type MediaInteractionCallback = (p: MediaInteractionParam) => void
 
 export type FocusCallback = (data: FocusRecord['data']) => void
-
-export type VisualViewportResizeCallback = (data: VisualViewportRecord['data']) => void
 
 export type ListenerHandler = () => void
 export type HookResetter = () => void
