+++ packages/rum/src/types.ts
@@ -28,13 +28,7 @@ export type CreationReason =
   | 'before_unload'
   | 'visibility_hidden'
 
-export type RawRecord =
-  | FullSnapshotRecord
-  | IncrementalSnapshotRecord
-  | MetaRecord
-  | FocusRecord
-  | ViewEndRecord
-  | VisualViewportRecord
+export type RawRecord = FullSnapshotRecord | IncrementalSnapshotRecord | MetaRecord | FocusRecord | ViewEndRecord
 
 export type Record = RawRecord & {
   timestamp: number
@@ -47,7 +41,6 @@ export const RecordType = {
   Meta: 4,
   Focus: 6,
   ViewEnd: 7,
-  VisualViewport: 8,
 } as const
 
 export type RecordType = typeof RecordType[keyof typeof RecordType]
@@ -86,17 +79,4 @@ export interface FocusRecord {
 
 export interface ViewEndRecord {
   type: typeof RecordType.ViewEnd
-}
-
-export interface VisualViewportRecord {
-  type: typeof RecordType.VisualViewport
-  data: {
-    scale: number
-    offsetLeft: number
-    offsetTop: number
-    pageLeft: number
-    pageTop: number
-    height: number
-    width: number
-  }
 }
